+++ src/components/dashboard.competition.increment.vue
@@ -5,7 +5,7 @@
   >
     <div class="px-4 py-5 sm:p-6">
       <dt class="text-sm font-medium text-gray-500 truncate">
-        {{ title }}
+        {{ logger.title }}
       </dt>
       <dd class="mt-1 text-3xl font-semibold text-gray-900">
         {{ logger.count }}
@@ -24,11 +24,6 @@ export default {
     };
   },
   props: ["logger"],
-  computed: {
-    title() {
-      return this.$store.state.keys[this.logger.id].title;
-    },
-  },
   methods: {
     onClick() {
       clearTimeout(this.timeout);
