+++ src/data.js
@@ -1,0 +1,65 @@
+
+export const competitionsJSON = [
+  {
+    id: 'dhikr',
+    title: 'Tasbih',
+    counters: {
+      dhikr_1: {
+        title: 'SubhanAllah'
+      },
+      dhikr_2: {
+        title: 'Alhamdulillah'
+      },
+      dhikr_3: {
+        title: 'Allahu Akbar'
+      },
+      dhikr_4: {
+        title: 'la ilaha ilAllah'
+      },
+    }
+  },
+  {
+    id: 'mindful',
+    title: 'Mindful Minutes',
+    counters: {
+      mindful_1: {
+        title: 'Listen to Quran'
+      }
+    }
+  },
+  {
+    id: 'charity',
+    title: 'Charity',
+    counters: {
+    }
+  },
+  {
+    id: 'fitness',
+    title: 'Fitness',
+    counters: {
+    }
+  }
+]
+
+export const competitionKeys = {
+  dhikr_1: {
+    competition: 'dhikr',
+    title: 'SubhanAllah'
+  },
+  dhikr_2: {
+    competition: 'dhikr',
+    title: 'Alhamdulillah'
+  },
+  dhikr_3: {
+    competition: 'dhikr',
+    title: 'Allahu Akbar'
+  },
+  dhikr_4: {
+    competition: 'dhikr',
+    title: 'la ilaha ilAllah'
+  },
+  mindful_1: {
+    competition: 'mindful',
+    title: 'Listen to Quran'
+  }
+}
