+++ src/main.js
@@ -44,18 +44,6 @@ new Vue({
 }).$mount('#app')
 
 
-// db.collection("users").add({
-//   first: "Ada",
-//   last: "Lovelace",
-//   born: 1815
-// })
-// .then((docRef) => {
-//   console.log("Document written with ID: ", docRef.id);
-// })
-// .catch((error) => {
-//   console.error("Error adding document: ", error);
-// });
-
 firebase.auth().onAuthStateChanged(function(user) {
   if (user) {
     store.dispatch('setUser', user);
