+++ src/state/Tab.js
@@ -1,0 +1,19 @@
+export default {
+  state: () => {
+    return {
+      activeTabId: 'dhikr',
+    }
+  },
+  mutations: {
+    SET_ACTIVE_ID(state, data) {
+      state.activeTabId = data;
+    },
+  },
+  actions: {
+    setActiveTab({commit}, id) {
+      commit('SET_ACTIVE_ID', id);
+    },
+  },
+  getters: {
+  },
+}
